+++ wizards/com/sun/star/wizards/ui/UnoDialog.java
@@ -229,8 +229,8 @@ public class UnoDialog implements EventNames
         XControl xControl2 = xDlgContainer.getControl(ControlName);
         XView xView = UnoRuntime.queryInterface(XView.class, xControl2);
         Size aSize = xView.getSize();
-        double dblMAPWidth = (double) (((Integer) Helper.getUnoPropertyValue(xControl2.getModel(), PropertyNames.PROPERTY_WIDTH)).intValue());
-        return (((double) (aSize.Width)) / dblMAPWidth);
+        double dblMAPWidth = (((Integer) Helper.getUnoPropertyValue(xControl2.getModel(), PropertyNames.PROPERTY_WIDTH)).intValue());
+        return ((aSize.Width) / dblMAPWidth);
     }
 
     public Size getpreferredLabelSize(String LabelName, String sLabel)
